Switched --privileged to :rw
Instead of using the --privileged flag and giving the container what
amounts to root access on the host, the :rw at the end of a volume bind
command, i.e. `-v /host_dir/:/container/dir/:rw`, gives the container
read/write on only that directory. This solves our previous permissions
issue in a much more elegant and security conscious way.
https://docs.docker.com/reference/run/#volume-shared-filesystems